requirements.txt
flask
tensorflow
pillow
numpy
opencv-python
gunicorn

gdown